move ocex commands to seperate file
